+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/base/models/WebhookConfig.kt
@@ -3,6 +3,8 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.base.models
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.toInputFile
+import com.github.insanusmokrassar.TelegramBotAPI.updateshandlers.UpdatesFilter
+import com.github.insanusmokrassar.TelegramBotAPI.updateshandlers.webhook.WebhookPrivateKeyConfig
 import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.*
 import io.ktor.server.cio.CIO
 import kotlinx.coroutines.CoroutineScope