+++ CHANGELOG.md
@@ -394,3 +394,4 @@ with new `PluginSerializer`
 ## 1.4.0
 
 * `PostsSchedulesTable` channels now are private and has no longer accessible
+* `PostsMessagesTable` now is using flows for output broadcasting and non-conflated broadcast channels inside