+++ CHANGELOG.md
@@ -396,3 +396,4 @@ with new `PluginSerializer`
 * `PostsSchedulesTable` channels now are private and has no longer accessible
 * `PostsMessagesTable` now is using flows for output broadcasting and non-conflated broadcast channels inside
 * Remove `Launch` deprecations
+* Actualize `Config`